requirements.txt
streamlit>=1.32.0
google-generativeai
reportlab
pillow
python-dotenv
requests
openpyxl
ezdxf
openai
